+++ CHANGELOG.md
@@ -16,8 +16,12 @@ Mapbox welcomes participation and contributions from everyone.
 
 ## 11.26.0 - 09 July, 2026
 
+### Features ✨ and improvements 🏁
 * Animate verticalFov along with other camera parameters.
 * Introduce new experimental `ViewAnnotationManager.viewAnnotationAvoidRegions` which allows to specify rectangular screen regions that view annotations should avoid. View annotations opt in to this behaviour via the new experimental `ViewAnnotation.enableAvoidRegions` option.
+
+### Bug fixes 🐞
+* [IndoorSelectorView] Fix scroll position not scrolling to the selected floor when switching to a different building.
 
 ## 11.26.0-rc.1 - 29 June, 2026